refactor: create a base class for view model tests and move the redundant setup code in it
