+++ app/src/test/java/com/simpleapp/challenge/ui/login/LoginViewModelTest.kt
@@ -1,22 +1,17 @@
 package com.simpleapp.challenge.ui.login
 
-import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.viewModelScope
 import com.simpleapp.challenge.data.model.User
 import com.simpleapp.challenge.data.repository.AuthRepository
+import com.simpleapp.challenge.ui.base.BaseViewModelTest
 import com.simpleapp.challenge.ui.login.LoginViewModel.Event
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.test.TestCoroutineDispatcher
-import kotlinx.coroutines.test.resetMain
-import kotlinx.coroutines.test.setMain
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mock
@@ -25,13 +20,7 @@ import org.mockito.Mockito.doReturn
 import org.mockito.junit.MockitoJUnitRunner
 
 @RunWith(MockitoJUnitRunner::class)
-class LoginViewModelTest {
-
-  @ExperimentalCoroutinesApi
-  val dispatcher = TestCoroutineDispatcher()
-
-  @get:Rule
-  val rule = InstantTaskExecutorRule()
+class LoginViewModelTest: BaseViewModelTest() {
 
   @Mock
   private lateinit var mockAuthRepository: AuthRepository
@@ -42,7 +31,6 @@ class LoginViewModelTest {
   @ExperimentalCoroutinesApi
   @Before
   fun setup() {
-    Dispatchers.setMain(dispatcher)
     vm = LoginViewModel(mockAuthRepository)
     vm.viewModelScope.launch {
       vm.eventsFlow.collect {
@@ -54,7 +42,6 @@ class LoginViewModelTest {
   @ExperimentalCoroutinesApi
   @After
   fun tearDown() {
-    Dispatchers.resetMain()
     eventsList.clear()
   }
